agenteazy-skills/jwt-decode/requirements.txt
pyjwt
fastapi
uvicorn
